fix missing config_nlu_template.yml

+++ rasa-nlu-entrypoint.sh
@@ -5,8 +5,11 @@ set -e
 apt-get update
 apt-get install -y gettext-base
 
-pip install -U "https://anaconda.org/intel/tensorflow/1.10.0/download/tensorflow-1.10.0-cp36-cp36m-linux_x86_64.whl-avx2.whl"
+wget -N "https://anaconda.org/intel/tensorflow/1.10.0/download/tensorflow-1.10.0-cp36-cp36m-linux_x86_64.whl-avx2.whl"
+pip install -U "tensorflow-1.10.0-cp36-cp36m-linux_x86_64.whl-avx2.whl"
 
-envsubst < /app/config_nlu_template.yml > /app/config_nlu.yml
+if [ -f /app/config_nlu_template.yml ]; then
+  envsubst < /app/config_nlu_template.yml > /app/config_nlu.yml
+fi
 
 exec "$@"
